Hoist fairminter fee to int; forward block_index to issuance fee gates
Two findings from fee correctness audit, both consistency cleanups:

M-1: fairminter.py:595 computed `fee = 0.5 * config.UNIT` (Python
  float = 50000000.0) and stored that float in issuances.fee_paid.
  The debit at line 705 used int(fee) so balance accounting was
  exact, but the schema column is INTEGER and other fee-paying paths
  (issuance, dividend) all hoist to int at compute time. SUM(REAL)
  in xcp_destroyed() drifts after ~9e15 fairminters (irrelevant in
  practice but breaks the int invariant). Hoist to int at compute.

L-1: issuance.validate's protocol.enabled() calls for the fee gates
  (pay_only_first_issuance, enable_rematch, numeric_asset_names,
  subassets, issuance_fee_update_*) lacked `block_index=block_index`,
  falling back to CurrentState. Same footgun pattern that
  2f5452967 fixed elsewhere. Forward the explicit block_index for
  consistent historical re-validation.

Found by fee correctness audit pass.

+++ counterparty-core/counterpartycore/lib/messages/fairminter.py
@@ -592,7 +592,11 @@ def parse(db, tx, message):
     else:
         # only new named assets have fees
         if existing_asset is None and not asset.startswith("A"):
-            fee = 0.5 * config.UNIT
+            # int (not float): issuances.fee_paid is INTEGER, debit uses int(fee).
+            # Other fee-paying messages (issuance, dividend) all hoist to int
+            # at compute time -- match the convention so SUM over REAL doesn't
+            # drift after enough fairminters and so the schema stays coherent.
+            fee = int(0.5 * config.UNIT)
 
     # we only premint if the faireminter is open and soft cap reached,
     # otherwise we will do it at opening (`start_block`) or when
